split api.ts
