+++ dashi/src/lib/api/types.ts
@@ -1,0 +1,4 @@
+export interface ApiOptions {
+  serverUrl?: string;
+  endpointName?: string;
+}
